requirements/prod.txt
asgiref==3.6.0
cffi==1.15.1
cryptography==39.0.2
Django==4.1.7
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-js-asset==2.0.0
django-model-utils==4.3.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
Pillow==9.4.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pytz==2022.7.1
sqlparse==0.4.3
tzdata==2022.7
Unipath==1.1